Move binding related function into lib/binding.ts
